feat: 공용 notifications 관련 로직 분리 및 구조화
