+++ src/basic/shared/index.ts
@@ -1,0 +1,4 @@
+export * from "./components";
+export * from "./hooks";
+export * from "./types";
+export * from "./utils";
