+++ src/basic/shared/types/notification.ts
@@ -1,0 +1,5 @@
+export interface Notification {
+  id: string;
+  message: string;
+  type: "error" | "success" | "warning";
+}
