+++ src/types.ts
@@ -1,3 +1,4 @@
 export type * from "./basic/domains/cart";
 export type * from "./basic/domains/coupon";
 export type * from "./basic/domains/product";
+export type * from "./basic/shared";
